requirements.txt

pandas
matplotlib
jinja2
python-dotenv
reportlab
schedule
